Explicitly set Psych gem dependency and version
The original settingslogic repo does not explicitly set a Psych dependency and
instead expects a dependency to exist by proxy. This can cause issues when that
proxy gem upgrades Psych to a version not compatible with settingslogic. This fork
and commit explicitly sets the Psych gem dependency and locks it to version 3.1.0